+++ src/test/java/io/github/astrapi69/xml/xpath/NodeExtensionsTest.java
@@ -28,22 +28,25 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.List;
 
 import javax.xml.parsers.ParserConfigurationException;
 import javax.xml.transform.TransformerException;
 import javax.xml.xpath.XPathExpressionException;
 
 import org.junit.jupiter.api.Test;
+import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
 import org.xml.sax.SAXException;
 
+import io.github.astrapi69.collections.list.ListFactory;
 import io.github.astrapi69.file.search.PathFinder;
 
 public class NodeExtensionsTest
 {
 
 	/**
-	 * Test method for {@link NodeExtensions#toString(NodeList, int)}
+	 * Test method for {@link NodeExtensions#toXmlString(NodeList, int)}
 	 *
 	 * @throws XPathExpressionException
 	 *             the x path expression exception
@@ -67,10 +70,42 @@ public class NodeExtensionsTest
 		xml = PathFinder.getRelativePath(PathFinder.getSrcTestResourcesDir(), "test-xml.xml");
 		xpathExpression = "/Customers/Customer";
 		nodeList = XPathExtensions.getNodeList(xml, xpathExpression);
-		actual = NodeExtensions.toString(nodeList, 0);
+		actual = NodeExtensions.toXmlString(nodeList, 0);
 		expected = "<?xml version=\"1.0\" encoding=\"UTF-8\"?><Customer id=\"1\">\n"
 			+ "\t\t<age>34</age>\n" + "\t\t<name>John</name>\n" + "\t\t<gender>Male</gender>\n"
 			+ "\t\t<role>Cpp Developer</role>\n" + "\t</Customer>";
 		assertEquals(expected, actual);
 	}
+
+	/**
+	 * Test method for {@link NodeExtensions#toXmlString(NodeList, int)}
+	 *
+	 * @throws XPathExpressionException
+	 *             the x path expression exception
+	 * @throws ParserConfigurationException
+	 *             the parser configuration exception
+	 * @throws SAXException
+	 *             the sAX exception
+	 * @throws IOException
+	 *             Signals that an I/O exception has occurred.
+	 */
+	@Test
+	public void testGetTextContentValues() throws XPathExpressionException,
+		ParserConfigurationException, SAXException, IOException, TransformerException
+	{
+		List<String> actual;
+		List<String> expected;
+		NodeList nodeList;
+		File xml;
+		String xpathExpression;
+
+		xml = PathFinder.getRelativePath(PathFinder.getSrcTestResourcesDir(), "company.xml");
+		xpathExpression = "/Company";
+		nodeList = XPathExtensions.getNodeList(xml, xpathExpression);
+		Node item = nodeList.item(0);
+		actual = NodeExtensions.getTextContentValues(item, "/Company/Customers/Customer/age");
+		expected = ListFactory.newArrayList("34", "32", "20", "28");
+		assertEquals(expected, actual);
+	}
+
 }